+++ src/audio/audio.h
@@ -41,7 +41,7 @@ public:
 	void bgmPlay(const char *filename,
 	             int volume = 100,
 	             int pitch = 100,
-	             float pos = 0,
+	             double pos = 0,
                  int track = -127);
 	void bgmStop(int track = -127);
 	void bgmFade(int time, int track = -127);
@@ -51,7 +51,7 @@ public:
 	void bgsPlay(const char *filename,
 	             int volume = 100,
 	             int pitch = 100,
-	             float pos = 0);
+	             double pos = 0);
 	void bgsStop();
 	void bgsFade(int time);
 
@@ -67,8 +67,8 @@ public:
 	void seStop();
 
 	void setupMidi();
-	float bgmPos(int track = 0);
-	float bgsPos();
+	double bgmPos(int track = 0);
+	double bgsPos();
 
 	void reset();
