+++ src/audio/vorbissource.cpp
@@ -156,7 +156,7 @@ struct VorbisSource : ALDataSource
 		return info.rate;
 	}
 
-	void seekToOffset(float seconds)
+	void seekToOffset(double seconds)
 	{
 		if (seconds <= 0)
 		{
@@ -164,6 +164,7 @@ struct VorbisSource : ALDataSource
 			currentFrame = 0;
 		}
 
+		// TODO: We're flooring here when we probably should be rounding.
 		currentFrame = seconds * info.rate;
 
 		if (loop.valid && currentFrame > loop.end)
